refactor(mcp): split mcp_handlers.rs into per-domain modules
Decompose the 1281-LOC src/server/mcp_handlers.rs into a directory module
of 7 focused submodules + shared tests, all within the project's 200-500
LOC target documented in docs/how-to/contribute.md.

  mod.rs            128  RPC dispatcher + matrix report
  builtin_tools.rs  210  tools/list catalog
  config.rs         195  read/apply/parse helpers
  configure.rs      203  configure + autotune handlers
  control_bridge.rs  66  MCP -> RPC bridge (keys/tools/hit/pledge)
  hint.rs            40  one-shot complexity hint
  wizard.rs         191  wizard get_config / set_section / run_doctor
  tests.rs          378  unit tests (relocated, 81/81 passing)

Public API preserved: handle_mcp_rpc and handle_matrix_report keep their
module path for src/server/mod.rs consumers; the rest are re-exported via
pub use to remain reachable.

Closes Tier-2 item #12 of the cli-cycle Phoenix triage.
